Use named lifetime for socket path
The current code uses a `'static` lifetime for the string passed into `with_socket_path` and
`with_name` which prevents a client from passing the socket path which is not statically
defined within the client's code.  By using an expliclt lifetime of `'a` it makes it so that
the client is in control over the value passed into the library.

Signed-off-by: Anthony Oteri <anthony.oteri@gmail.com>